refactor(api): remove deprecated menu system and related endpoints
Remove legacy menu router, MenuDocument, and associated database
migration code. This includes:
- Remove menu router from API main application
- Remove menu-related calibration scheduling endpoints
- Remove MenuDocument from database initialization
- Remove v1.0.16 to v1.0.17 menu migration code

The menu functionality has been superseded by the flow-based
calibration system.
