+++ src/qdash/db/init/__init__.py
@@ -2,7 +2,6 @@
 
 from qdash.db.init.chip import init_chip_document
 from qdash.db.init.coupling import init_coupling_document
-from qdash.db.init.menu import init_menu
 from qdash.db.init.qubit import init_qubit_document
 from qdash.db.init.task import init_task_document
 
@@ -10,6 +9,5 @@ __all__ = [
     "init_chip_document",
     "init_coupling_document",
     "init_qubit_document",
-    "init_menu",
     "init_task_document",
 ]
