--- ui/src/app/menu/editor/DeleteTaskConfirmModal.tsx
@@ -1,56 +1,0 @@
-"use client";
-
-import { BsPlus } from "react-icons/bs";
-
-export function DeleteTaskConfirmModal({
-  taskName,
-  onConfirm,
-  onClose,
-}: {
-  taskName: string;
-  onConfirm: () => void;
-  onClose: () => void;
-}) {
-  return (
-    <div
-      className="fixed inset-0 bg-black/60 flex items-center justify-center z-50 backdrop-blur-sm p-4"
-      onClick={onClose}
-    >
-      <div
-        className="bg-base-100 rounded-xl w-full max-w-lg overflow-hidden flex flex-col shadow-2xl"
-        onClick={(e) => e.stopPropagation()}
-      >
-        <div className="px-6 py-4 border-b border-base-300 flex items-center justify-between bg-base-100/80 backdrop-blur supports-[backdrop-filter]:bg-base-100/60">
-          <div>
-            <h2 className="text-2xl font-bold">Delete Task</h2>
-            <p className="text-base-content/70 mt-1">
-              This action cannot be undone
-            </p>
-          </div>
-          <button
-            onClick={onClose}
-            className="btn btn-ghost btn-sm btn-square hover:rotate-90 transition-transform"
-          >
-            <BsPlus className="text-xl rotate-45" />
-          </button>
-        </div>
-
-        <div className="p-6">
-          <p>
-            Are you sure you want to delete the task{" "}
-            <span className="font-semibold">{taskName}</span>?
-          </p>
-        </div>
-
-        <div className="px-6 py-4 border-t border-base-300 flex justify-end gap-2">
-          <button className="btn btn-ghost" onClick={onClose}>
-            Cancel
-          </button>
-          <button className="btn btn-error" onClick={onConfirm}>
-            Delete
-          </button>
-        </div>
-      </div>
-    </div>
-  );
-}
